example: use uint8_t for all maple send_buf/recv_buf

+++ example/clipping_textured.cpp
@@ -36,8 +36,8 @@
 
 static ft0::data_transfer::data_format data[4];
 
-uint32_t send_buf[1024] __attribute__((aligned(32)));
-uint32_t recv_buf[1024] __attribute__((aligned(32)));
+uint8_t send_buf[1024] __attribute__((aligned(32)));
+uint8_t recv_buf[1024] __attribute__((aligned(32)));
 
 void do_get_condition()
 {
